+++ examples/two-spheres.py
@@ -8,24 +8,25 @@ import matplotlib.pyplot as plt
 
 # Parameters
 mesh_size: float = 0.012  # (m)
-mu_r: float = 1.5
+mu_r: float = 2.5
 b_external_magnitude: float = 5.0  # (T)
+radius: float = 0.05  # (m)
+distance: float = 0.3  # (m)
 
 # Create two spheres, each of 0.1m diameter separated by a CL-CL distance of 0.3m
 # The CAD file is for a sphere centered at (0.0, 0.0, 0.0)
-radius: float = 0.05  # (m)
-distance: float = 0.3  # (m)
+
 upper_sphere: Mesh = Mesh.from_step("tests/data/sphere.stp", mesh_size)
-lower_sphere: Mesh = Mesh(upper_sphere.nodes.copy(), upper_sphere.connectivity.copy())
-upper_sphere.nodes[:, 2] += 0.15
-lower_sphere.nodes[:, 2] -= 0.15
+upper_sphere.nodes[:, 2] += distance / 2.0
+lower_sphere: Mesh = Mesh.from_step("tests/data/sphere.stp", mesh_size)
+lower_sphere.nodes[:, 2] -= distance / 2.0
 
 # Combine the meshes together so that we can compute field interactions
-mesh = lower_sphere.append(upper_sphere)
-# mesh.plot()
+combined_mesh = lower_sphere.append(upper_sphere)
+# combined_mesh.plot("docs/figs/example_two_spheres_mesh.svg")
 
 # Compute a constant external field
-h_external = np.zeros(mesh.centroids.shape)
+h_external = np.zeros(combined_mesh.centroids.shape)
 h_external[:, 2] = b_external_magnitude / oersted.MU0
 
 # Assign a material
@@ -33,10 +34,10 @@ material = oersted.LinearMaterial(mu_r)
 
 # Create a solver
 # solver = oersted.OctreeSolver(leaf_threshold=1, alpha=0.9, tol=1e0)
-solver = oersted.DirectSolver(alpha=0.9)
+solver = oersted.DirectSolver(alpha=0.5, tol=1e0)
 
 # Demagnetization solve on both spheres
-M, _ = oersted.demag_solve(mesh, material, h_external, solver)
+M, _ = oersted.demag_solve(combined_mesh, material, h_external, solver)
 
 # Select only the magnetization field in the appropriate sphere
 M_lower = M[: lower_sphere.num_elems]
@@ -61,8 +62,14 @@ forces_lower = oersted.kelvin_forces(lower_sphere, M_lower, MU0 * h_field_nodes_
 force_upper = np.sum(forces_upper, axis=0)
 force_lower = np.sum(forces_lower, axis=0)
 
-print(f"Force on upper sphere: {force_upper}")
-print(f"Force on lower sphere: {force_lower}")
+print(
+    f"Force on upper sphere: \
+        ({force_upper[0]:.1f},{force_upper[1]:.1f},{force_upper[2]:.1f}) N "
+)
+print(
+    f"Force on lower sphere: \
+        ({force_lower[0]:.1f},{force_lower[1]:.1f},{force_lower[2]:.1f}) N"
+)
 
 # Analytical solution; see: <https://en.wikipedia.org/wiki/Force_between_magnets#Magnetic_dipole-dipole_interaction>
 N: float = 1.0 / 3.0  # demag factor for sphere
@@ -72,13 +79,15 @@ H_analytical: float = H_ext - N * M_analytical
 B_analytical = oersted.MU0 * (H_ext + (1.0 - N) * M_analytical)
 m = (4.0 / 3.0) * np.pi * (radius**3) * M_analytical
 F_analytical = 3 * MU0 * m * m / (2 * np.pi * distance**4)
+print(f"Analytical M field: {M_analytical:.3f} A/m")
+print(f"Analytical magnetic moment: {m:.3f} A-m^2")
 print(f"Analytical force: {F_analytical}")
 
 # Compute the fields along a line going through the spheres
 n = 1000
 targets = np.zeros((n, 3))
 targets[:, 2] = np.linspace(-0.2, 0.2, n)
-htargets = oersted.h_mag(mesh, M, targets, solver)
+htargets = oersted.h_mag(combined_mesh, M, targets, solver)
 fig, ax = plt.subplots()
 ax.plot(targets[:, 2], htargets[:, 2])
 ax.set_xlabel("Distance Along Z-Axis (m)")