+++ CHANGELOG.md
@@ -91,4 +91,5 @@
 * Cleaned up docstrings
 * Fixed Rust code to provide `h_current_point_direct_parallel` instead of `bfield_direct`, 
 bringing it in-line with the tetrahedron functions
-* Organized docs and updated
+* Organized and updated docs
+* Added two spheres example writeup